Update project configuration and files

+++ backend/src/server/services/cliente/clienteControllers.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import Cliente from './clienteModels';
 import Pessoa from '../pessoa/pessoaModels';
+import Representante from '../representante/representanteModels';
 
 // Defina o controlador para o modelo 'Cliente'
 class ClienteController {
@@ -22,6 +23,34 @@ class ClienteController {
         }
     }
 
+    public async update(req: Request, res: Response): Promise<Response> {
+        const { id } = req.params;
+
+        try {
+            const cliente = await Cliente.findByPk(id);
+
+            if (!cliente) {
+                return res.status(404).json({ error: 'Cliente não encontrado' });
+            }
+
+            // Verifica se a Pessoa já existe pelo identificador (identificacao)
+            const identificacao = req.body.identificacao;
+            let pessoa = await Pessoa.findOne({ where: { identificacao } });
+            // Se a Pessoa não existe, retorna erro 404
+            if (!pessoa) {
+                return res.status(404).json({ error: 'Cliente não exite' });
+            }else{
+                // Atualiza a Pessoa
+                await Pessoa.update(req.body, { where: { id: pessoa.id } });
+            }
+            // Atualiza o Cliente
+            await Cliente.update(req.body, { where: { id } });
+
+            return res.status(200).json(cliente);
+        } catch (error: any) {
+            return res.status(500).json({ error: 'Erro interno do servidor' });
+        }
+    }
 
     public async getAll(req: Request, res: Response): Promise<Response> {
         try {
@@ -41,8 +70,6 @@ class ClienteController {
         }
     }
 
-    
-
     public async getById(req: Request, res: Response): Promise<Response> {
         const { id } = req.params;
 
@@ -59,33 +86,25 @@ class ClienteController {
         }
     }
 
-    public async update(req: Request, res: Response): Promise<Response> {
-        const { id } = req.params;
-
-        try {
-            const [updatedRows] = await Cliente.update(req.body, { where: { id } });
-
-            if (updatedRows === 0) {
-                return res.status(404).json({ error: 'Cliente não encontrado' });
-            }
-
-            const cliente = await Cliente.findByPk(id, { include: Pessoa });
-
-            return res.status(200).json(cliente);
-        } catch (error: any) {
-            return res.status(500).json({ error: 'Erro interno do servidor' });
-        }
-    }
-
     public async delete(req: Request, res: Response): Promise<Response> {
         const { id } = req.params;
 
         try {
-            const deletedRows = await Cliente.destroy({ where: { id } });
 
-            if (deletedRows === 0) {
+            // Verifica se o Cliente existe
+            const cliente = await Cliente.findByPk(id);
+            if (!cliente) {
                 return res.status(404).json({ error: 'Cliente não encontrado' });
             }
+
+            // Verifica se há um Representante associado ao Cliente
+            const representante = await Representante.findOne({ where: { id_cliente: id } });
+            if (representante) {
+                return res.status(400).json({ error: 'Não é possível remover o Cliente com um Representante associado' });
+            }
+
+            // Remove o Cliente
+            await Cliente.destroy({ where: { id } });
 
             return res.status(204).send();
         } catch (error: any) {
